Operator precedence parser
In this commit I have tried interconnect scanner with operator-precedence parser. There was a problem with getting tokens and I have decided to try simple getchar function and get limited tokens just for '5' as identifier and '+' as operator. It worked. Operator precedence parser generated postfix notation of expression. However, there was a problem with EOF. Still need to solve problem with EOF to detect it correctly.

TODO:
	Test scanner in order to detect any bugs